feat(aimcp): implement MCP client functionality and enhance related features
- Add MCPClientRegistry to manage MCP clients
- Implement CommandEnvStrategy interface and its factory
- Add support for Java, Node.js, and Python command environments
- Enhance AiMcpService with file upload and tool specification parsing
-

+++ src/main/java/org/hzai/ai/aimcptools/entity/AiMcpTools.java
@@ -44,7 +44,7 @@ public class AiMcpTools extends PanacheEntityBase {
 	@Column(columnDefinition = "INT DEFAULT 0")
 	private Integer isDeleted;
 
-	private Integer mcpId;
+	private Long mcpId;
 
 
 }
